Updated DB Schema (#10)
* Updated DB Schema

This includes the new 'seatingLayoutSchema', an update to the 'attendence' property in courseSchema, and some other minor updates

* Update attendanceSchema.js (#11)

fixed so that the email is unique for users

* Update attendanceSchema.js (#16)

These are updates to the changes we discussed. I ended up keeping `seating_arrangement`, `course_type`, and `mandatory` in the `attendance` object due to my reasons from the pr. If you feel we should talk through it more and change it, please let me now.

* Update attendanceSchema.js

* Update attendanceSchema.js