Update attendanceSchema.js
While working on the api endpoint, we realized we forgot to add a `created_by` property to `seating_layout` so that teachers could access previous layouts THEY created. Plus I corrected a few minor spelling mistakes This is just the addition of one line and is a quick patch but still needs to be approved. SORRY FOR THE SPAM